+++ scrapy/spiders/feed.py
@@ -5,7 +5,7 @@ for scraping from an XML feed.
 See documentation in docs/topics/spiders.rst
 """
 from scrapy.spiders import Spider
-from scrapy.utils.iterators import xmliter, csviter
+from scrapy.utils.iterators import csviter, xmliter_lxml
 from scrapy.utils.spider import iterate_spider_output
 from scrapy.selector import Selector
 from scrapy.exceptions import NotConfigured, NotSupported
@@ -82,7 +82,7 @@ class XMLFeedSpider(Spider):
         return self.parse_nodes(response, nodes)
 
     def _iternodes(self, response):
-        for node in xmliter(response, self.itertag):
+        for node in xmliter_lxml(response, self.itertag):
             self._register_namespaces(node)
             yield node